Доработки по селекторам
