+++ project/src/components/card/card.tsx
@@ -1,4 +1,6 @@
 import { Link } from 'react-router-dom';
+import { useAppDispatch } from '../../hooks/hooks';
+import { Action } from '../../store/action';
 import Film from '../../types/film';
 import Videoplayer from '../videoplayer/videoplayer';
 
@@ -7,11 +9,18 @@ type CardProps = {
 };
 
 function Card({film} : CardProps):JSX.Element{
+  const dispatch = useAppDispatch();
   return (
     <article className="small-film-card catalog__films-card">
       <Videoplayer src={film.videoSrc} imgSrc='img/snatch.jpg'/>
       <h3 className="small-film-card__title">
-        <Link to={`/films/${film.id}`} className="small-film-card__link">{film.name}</Link>
+        <Link to={`/films/${film.id}`} className="small-film-card__link"
+          onClick={() => {
+            dispatch(Action.FILM.SET_CURRENT({currentFilm: film.id}));
+          }}
+        >
+          {film.name}
+        </Link>
       </h3>
     </article>
   );
