+++ project/src/pages/film-screen/film-screen.tsx
@@ -1,24 +1,16 @@
-import { Link, useParams } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import FilmsList from '../../components/films-list/films-list';
+import Header from '../../components/header/header';
 import Tabs from '../../components/tabs/tabs';
-import { useAppDispatch, useAppSelector } from '../../hooks/hooks';
-import { Action } from '../../store/action';
-import { selectCurrentFilm, selectFilteredFilms } from '../../store/selectors';
+import { useAppSelector } from '../../hooks/hooks';
+import { getFilteredFilms } from '../../store/selectors';
 
 function FilmScreen() : JSX.Element {
   // eslint-disable-next-line no-console
   console.log('film');
 
-  const dispatch = useAppDispatch();
-  const currentFilmId = Number(useParams().id);
-  let currentFilm = selectCurrentFilm(useAppSelector((state) => state));
-
-  if(currentFilm.id !== currentFilmId){
-    dispatch(Action.FILM.SET_CURRENT({currentFilm: currentFilmId}));
-  }
-
-  currentFilm = selectCurrentFilm(useAppSelector((state) => state));
-  const filteredFilms = selectFilteredFilms(useAppSelector((state) => state));
+  const {currentFilm} = useAppSelector((state) => state);
+  const filteredFilms = useAppSelector(getFilteredFilms);
 
   return (
     <>
@@ -27,26 +19,7 @@ function FilmScreen() : JSX.Element {
           <div className="film-card__bg">
             <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
           </div>
-          <h1 className="visually-hidden">WTW</h1>
-          <header className="page-header film-card__head">
-            <div className="logo">
-              <Link to='/' className="logo__link">
-                <span className="logo__letter logo__letter--1">W</span>
-                <span className="logo__letter logo__letter--2">T</span>
-                <span className="logo__letter logo__letter--3">W</span>
-              </Link>
-            </div>
-            <ul className="user-block">
-              <li className="user-block__item">
-                <div className="user-block__avatar">
-                  <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-                </div>
-              </li>
-              <li className="user-block__item">
-                <a className="user-block__link">Sign out</a>
-              </li>
-            </ul>
-          </header>
+          <Header />
           <div className="film-card__wrap">
             <div className="film-card__desc">
               <h2 className="film-card__title">{currentFilm.name}</h2>
